Update save config function
